tests: Move functions to libpod
