+++ tests/libpod/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "blackhole.go",
         "certs.go",
         "container.go",
+        "pod.go",
         "query.go",
         "render.go",
         "sort.go",
@@ -18,6 +19,7 @@ go_library(
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests/exec:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
+        "//tests/framework/matcher:go_default_library",
         "//tests/libregistry:go_default_library",
         "//tests/testsuite:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
