+++ tests/storage/reservation.go
@@ -89,7 +89,7 @@ var _ = SIGDescribe("[Serial]SCSI persistent reservation", Serial, func() {
 		libstorage.CreateFSPVC(pvc, testsuite.NamespacePrivileged, diskSize, nil)
 		// Create targetcli container
 		By("Create targetcli pod")
-		pod := tests.RunPodInNamespace(libpod.RenderTargetcliPod(podName, pvc), testsuite.NamespacePrivileged)
+		pod := libpod.RunPodInNamespace(libpod.RenderTargetcliPod(podName, pvc), testsuite.NamespacePrivileged)
 		node = pod.Spec.NodeName
 		// The vm-killer image is built with bazel and the /etc/ld.so.cache isn't built
 		// at the package installation. The targetcli utility relies on ctype python package that
